Libft

+++ Libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: dramirez <dramirez@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/27 21:15:45 by dramirez          #+#    #+#             */
-/*   Updated: 2022/09/27 21:15:45 by dramirez         ###   ########.fr       */
+/*   Updated: 2022/10/11 23:37:38 by dramirez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	contador;
+	size_t	contador;
 
 	contador = 0;
 	while (s[contador] != '\0')
